requirements.txt
flask==2.3.3
Pillow==10.0.1
google-generativeai==0.3.1
python-dotenv==1.0.0